# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Study Guide API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id?: number;
  clerkUserId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface CreditBalance {
  creditsRemaining: number;
  totalPurchased: number;
  freeCreditUsed?: boolean;
}

export interface UserStats {
  totalAnalyses: number;
  creditsUsed: number;
  subjectsAnalyzed: string[];
  creditsRemaining: number;
}

export type AnalysisSummaryCategory = typeof AnalysisSummaryCategory[keyof typeof AnalysisSummaryCategory];


export const AnalysisSummaryCategory = {
  school: 'school',
  college: 'college',
} as const;

export type AnalysisSummaryStatus = typeof AnalysisSummaryStatus[keyof typeof AnalysisSummaryStatus];


export const AnalysisSummaryStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface AnalysisSummary {
  id: number;
  category: AnalysisSummaryCategory;
  /** @nullable */
  classOrCourse?: string | null;
  /** @nullable */
  boardOrUniversity?: string | null;
  subject: string;
  /** @nullable */
  yearsAnalyzed?: number | null;
  status: AnalysisSummaryStatus;
  hasPdf?: boolean;
  createdAt: string;
}

export type AnalysisCategory = typeof AnalysisCategory[keyof typeof AnalysisCategory];


export const AnalysisCategory = {
  school: 'school',
  college: 'college',
} as const;

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type ChapterResultPriority = typeof ChapterResultPriority[keyof typeof ChapterResultPriority];


export const ChapterResultPriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface ChapterResult {
  chapter_name: string;
  frequency: number;
  marks_weightage: string;
  priority: ChapterResultPriority;
  /** @nullable */
  study_note?: string | null;
}

export interface AiAnalysisResult {
  subject: string;
  category: string;
  years_analyzed: number;
  chapters: ChapterResult[];
  overall_strategy_tip: string;
}

export interface Analysis {
  id: number;
  category: AnalysisCategory;
  /** @nullable */
  classOrCourse?: string | null;
  /** @nullable */
  boardOrUniversity?: string | null;
  subject: string;
  /** @nullable */
  yearsAnalyzed?: number | null;
  status: AnalysisStatus;
  /** @nullable */
  errorMessage?: string | null;
  aiResponse?: AiAnalysisResult;
  hasPdf?: boolean;
  createdAt: string;
}

export type AnalysisInputCategory = typeof AnalysisInputCategory[keyof typeof AnalysisInputCategory];


export const AnalysisInputCategory = {
  school: 'school',
  college: 'college',
} as const;

export interface AnalysisInput {
  category: AnalysisInputCategory;
  classOrCourse?: string;
  boardOrUniversity?: string;
  subject: string;
  filePaths: string[];
}

export interface PdfDownload {
  url: string;
}

export interface PaymentOrder {
  orderId: string;
  amount: number;
  currency: string;
  key: string;
}

export interface PaymentVerification {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface Payment {
  id: number;
  amount: number;
  /** @nullable */
  razorpayPaymentId?: string | null;
  status: PaymentStatus;
  createdAt: string;
}