# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Study Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisSummaryCategory } from './analysisSummaryCategory';
import type { AnalysisSummaryStatus } from './analysisSummaryStatus';

export interface AnalysisSummary {
  id: number;
  category: AnalysisSummaryCategory;
  /** @nullable */
  classOrCourse?: string | null;
  /** @nullable */
  boardOrUniversity?: string | null;
  subject: string;
  /** @nullable */
  yearsAnalyzed?: number | null;
  status: AnalysisSummaryStatus;
  hasPdf?: boolean;
  createdAt: Date;
}