# lib/api-zod/src/generated/types/chapterResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Study Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterResultPriority } from './chapterResultPriority';

export interface ChapterResult {
  chapter_name: string;
  frequency: number;
  marks_weightage: string;
  priority: ChapterResultPriority;
  /** @nullable */
  study_note?: string | null;
}