# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Study Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentStatus } from './paymentStatus';

export interface Payment {
  id: number;
  amount: number;
  /** @nullable */
  razorpayPaymentId?: string | null;
  status: PaymentStatus;
  createdAt: Date;
}