# lib/api-zod/src/generated/types/paymentVerification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Study Guide API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerification {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}